loss callback cost and total_cost hardcoded to ensure accurate accumulation in fp16

This should avoid the Inf in fp16 alexnet cost. I am using this scheme on the flex branch and it works, no need to replace the hacky accumulate

self.total_cost[:] = self.total_cost + self.cost.get_cost(x, t)

by something clean. Passes tests so apparently no issue with CPU runs that expect fp64

See merge request !140

+++ neon/callbacks/callbacks.py
@@ -485,7 +485,7 @@ class LossCallback(Callback):
         super(LossCallback, self).__init__(epoch_freq=epoch_freq)
         self.model = model
         self.eval_set = eval_set
-        self.loss = self.be.zeros((1, 1))
+        self.loss = self.be.zeros((1, 1), dtype=np.float32)
         self.callback_data = callback_data
 
     def on_train_begin(self, epochs):
